requirements.txt
opencv-contrib-python>=4.0.0
google-auth>=2.3.0
google-cloud-core>=2.0.0
google-cloud-speech>=2.0.0
pyaudio>=0.2.11
scipy>=1.7.0
pillow>=8.0.0
matplotlib>=3.4.0
numpy>=1.20.0